Overhaul the relationships between classes
Despite how many changes this commit does, it basically narrows
down to the changes of the responsibilities of the classes.

BackupItem doesn't add chain items into a ChainConfigurationFrame
instance. It only passes the name of the backup, and the
initialization of inner components is up to the
ChainConfigurationFrame instance.

ChainConfigurationFrame doesn't have methods to control elements on
the list to remove code duplication. So chain items get reference to
the ChainList to manipulate its elements.

All chain items outside ChainList are now replaced with chain rules,
which support null values. It's allowed to not set a source and/or a
destination in ChainConfigurationFrame, however, chain items don't
support null values, so crunches were used. Now ChainSet methods can
store unset values like it is an empty line and ChainItem can
explicitly convey to the users that the value wasn't set, but both
classes use null to specify unset values.